requirements.txt
apscheduler==3.10.0
beautifulsoup4==4.11.2
dbutils==3.0.2
jaydebeapi==1.2.3
JPype1==1.4.1
netCDF4==1.6.2
numpy==1.24.2
pymysql==1.0.2
requests==2.28.2
